+++ stoq/installer.py
@@ -65,9 +65,10 @@ class StoqPluginInstaller:
                 .replace('@', '/')
             )
             path = plugin_path.split('subdirectory=')[1]
-            requirements = f'{url}/{path}/requirements.txt'
+            requirements = f'{url}/master/{path}/requirements.txt'
             with NamedTemporaryFile() as temp_file:
                 response = requests.get(requirements)
+                response.raise_for_status()
                 if response.status_code == 200:
                     temp_file.write(response.content)
                     temp_file.flush()
@@ -82,6 +83,8 @@ class StoqPluginInstaller:
                             temp_file.name,
                         ]
                     )
+                else:
+                    print(f'Failed to install requirements from {requirements}')
         else:
             requirements = f'{plugin_path}/requirements.txt'
             if os.path.isfile(requirements):